requirements.txt
fastapi>=0.110
uvicorn>=0.27
openai>=1.0
pydantic>=2.0
python-dotenv>=1.0
python-multipart>=0.0.6
